requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.0.1
mysql-connector-python==8.1.0
bcrypt==4.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7 